requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
numpy>=1.26.0
httpx>=0.27.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
